Refactor invoke commands to test the guts.
The invoke commands involve reading from files and
writing to files, and that is one of the reasons they
are a pain to test because it would involve a lot of
setting up and tearing down.

In this patch, the API is modified so that each invoke
command calls internal functions that deal with python
objects (mostly pandas.DataFrames) that can be tested
in memory.